moved fetching data to hooks & added icons for active markers
